remove old base date picker
