--- client/packages/common/src/ui/components/inputs/DatePickers/BaseDatePickerInput/index.ts
@@ -1,1 +1,0 @@
-export * from './BaseDatePickerInput';
