+++ client/packages/common/src/ui/components/inputs/DatePickers/index.ts
@@ -1,3 +1,2 @@
-export * from './BaseDatePickerInput';
 export * from './DateTimePickerInput';
 export * from './ExpiryDateInput';
